moved feature constants to separate file
