+++ tabular/src/autogluon/tabular/data/utils.py
@@ -4,7 +4,7 @@ import math
 from pandas import DataFrame
 
 from autogluon.core.features.types import get_type_map_raw
-from autogluon.core.features.feature_metadata import R_INT, R_FLOAT, R_CATEGORY
+from autogluon.core.features.constants import R_INT, R_FLOAT, R_CATEGORY
 
 logger = logging.getLogger(__name__)
