+++ tabular/src/autogluon/tabular/features/__init__.py
@@ -1,2 +1,1 @@
-from autogluon.core.features.feature_metadata import *
 from .generators import *
