+++ tabular/src/autogluon/tabular/features/generators/astype.py
@@ -2,7 +2,8 @@ import logging
 
 from pandas import DataFrame, Series
 
-from autogluon.core.features.feature_metadata import FeatureMetadata, R_INT
+from autogluon.core.features.constants import R_INT
+from autogluon.core.features.feature_metadata import FeatureMetadata
 from autogluon.core.features.types import get_type_map_real
 
 from .abstract import AbstractFeatureGenerator
