+++ tabular/src/autogluon/tabular/features/generators/datetime.py
@@ -3,7 +3,7 @@ import logging
 import pandas as pd
 from pandas import DataFrame
 
-from autogluon.core.features.feature_metadata import R_DATETIME, S_DATETIME_AS_OBJECT
+from autogluon.core.features.constants import R_DATETIME, S_DATETIME_AS_OBJECT
 
 from .abstract import AbstractFeatureGenerator
