+++ tabular/src/autogluon/tabular/features/generators/fillna.py
@@ -3,7 +3,7 @@ import logging
 import numpy as np
 from pandas import DataFrame
 
-from autogluon.core.features.feature_metadata import R_OBJECT
+from autogluon.core.features.constants import R_OBJECT
 
 from .abstract import AbstractFeatureGenerator
