+++ tabular/src/autogluon/tabular/features/generators/label_encoder.py
@@ -3,7 +3,7 @@ import logging
 
 from pandas import DataFrame
 
-from autogluon.core.features.feature_metadata import R_CATEGORY, S_TEXT_AS_CATEGORY
+from autogluon.core.features.constants import R_CATEGORY, S_TEXT_AS_CATEGORY
 
 from .abstract import AbstractFeatureGenerator
