+++ tabular/src/autogluon/tabular/features/generators/text_ngram.py
@@ -7,7 +7,7 @@ import pandas as pd
 import psutil
 from pandas import DataFrame
 
-from autogluon.core.features.feature_metadata import S_TEXT, S_TEXT_NGRAM
+from autogluon.core.features.constants import S_TEXT, S_TEXT_NGRAM
 
 from .abstract import AbstractFeatureGenerator
 from ..vectorizers import get_ngram_freq, downscale_vectorizer, vectorizer_auto_ml_default
