+++ tabular/src/autogluon/tabular/models/ensemble/stacker_ensemble_model.py
@@ -7,7 +7,8 @@ from collections import defaultdict
 
 from autogluon.core.utils.utils import generate_kfold
 from autogluon.core.constants import MULTICLASS
-from autogluon.core.features.feature_metadata import FeatureMetadata, R_FLOAT, S_STACK
+from autogluon.core.features.constants import R_FLOAT, S_STACK
+from autogluon.core.features.feature_metadata import FeatureMetadata
 
 from autogluon.core.models import AbstractModel
 from .bagged_ensemble_model import BaggedEnsembleModel
