+++ tabular/src/autogluon/tabular/models/lgb/lgb_model.py
@@ -14,7 +14,7 @@ from autogluon.core.utils.savers import save_pkl
 from autogluon.core.utils import try_import_lightgbm
 from autogluon.core import Int, Space
 from autogluon.core.constants import BINARY, MULTICLASS, REGRESSION, SOFTCLASS
-from autogluon.core.features.feature_metadata import R_OBJECT
+from autogluon.core.features.constants import R_OBJECT
 
 from . import lgb_utils
 from .callbacks import early_stopping_custom
