+++ tabular/src/autogluon/tabular/models/tab_transformer/tab_transformer_model.py
@@ -8,7 +8,7 @@ import pandas as pd
 from tqdm import tqdm
 
 from autogluon.core.constants import BINARY, REGRESSION, MULTICLASS
-from autogluon.core.features.feature_metadata import R_OBJECT, S_TEXT_NGRAM, S_TEXT_AS_CATEGORY
+from autogluon.core.features.constants import R_OBJECT, S_TEXT_NGRAM, S_TEXT_AS_CATEGORY
 from autogluon.core.utils import try_import_torch
 from autogluon.core.utils.loaders import load_pkl
