+++ tabular/src/autogluon/tabular/models/text_prediction/text_prediction_v1_model.py
@@ -9,7 +9,7 @@ import random
 import numpy as np
 import pandas as pd
 
-from autogluon.core.features.feature_metadata import R_OBJECT, R_INT, R_FLOAT, R_CATEGORY, \
+from autogluon.core.features.constants import R_OBJECT, R_INT, R_FLOAT, R_CATEGORY, \
     S_TEXT_NGRAM, S_TEXT_AS_CATEGORY, S_TEXT_SPECIAL
 from autogluon.core.utils import get_cpu_count, get_gpu_count
 from autogluon.core.utils.exceptions import NoGPUError, NoValidFeatures
